Hi! Please pay attention
Hi! Amel, I finished this online class, but the weird thing is the successful badge showed in the middle of this class --lesson 3 instead of at the end, thus I did separated screenshots for each lesson with my portrait at the top right, as long as I finished one,  to confirm I completed this! Have a nice day!

Regards, Miaozi Yin